feature: make AnimatedLargeTopAppBar
